Run invoice processing on 1st of every month
Scheduling a job for the first day of the next month to process
is quite easy, the tricky part is to get subsequent invoice processing jobs done.
We avoid this issue by not scheduling 'processInvoices' job, but
rather 'dailyBillingCheck' that gets scheduled at a fixed rate every
day. The job checks if today is the first day of the month and runs
the invoice processing, otherwise does nothing.

* Add 'dailyBillingCheck' job that gets executed initially at the first day of the
  next month, then each day. Here we check if today is first day of the month and run
  'processInvoices'
* Expose executor arguments as 'scheduleJob' parameters. This way the scheduler remains
  generic and can be used for other purposes as well.